requirements.txt
numpy<2.0.0
torch
torchvision
torchaudio
diffusers
transformers
accelerate
pillow
opencv-python
streamlit
protobuf<=3.20
